Added GlobalGameLeaders.com Integration
When you win, or if you lose during the Platforming Stage it will now
report the result to globalgameleaders.com

I also did a little bit of tidy up, and fixing. The minigames should
hopefully now play the correct music (untested), and I made the Heart
decay a lot faster.